+++ integrations/github/src/api.ts
@@ -245,6 +245,11 @@ async function githubAPI<T>(
 
     const response = await requestGitHubAPI(context, credentials, url, options);
 
+    const isJSONResponse = response.headers.get('Content-Type')?.includes('application/json');
+    if (!isJSONResponse) {
+        return (await response.text()) as unknown as T;
+    }
+
     let data = await response.json();
 
     let paginatedListProperty = false;