+++ packages/api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gitbook/api
 
+## 0.28.0
+
+### Minor Changes
+
+-   4cb31e9: Bump @gitbook/api to use the latest openAPI spec
+
+## 0.27.0
+
+### Minor Changes
+
+-   46c9686: Allow passing a `userAgent` to the API client (`@gitbook/api`) and set a default user-agent for all integrations
+
 ## 0.26.0
 
 ### Minor Changes